+++ src/main/java/gyeongdan/article/article/dto/ArticleAllResponse.java
@@ -1,14 +1,11 @@
-package gyeongdan.article.dto;
+package gyeongdan.article.article.dto;
 
-import gyeongdan.article.domain.Article;
-import gyeongdan.article.domain.ArticleRelatedDocuments;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 
 import java.sql.Timestamp;
-import java.util.List;
 
 @Setter
 @Getter
