+++ src/main/java/gyeongdan/article/article/dto/PopularArticleResponse.java
@@ -1,4 +1,4 @@
-package gyeongdan.article.dto;
+package gyeongdan.article.article.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
