+++ src/main/java/gyeongdan/article/article/repository/ArticleJpaRepository.java
@@ -1,6 +1,6 @@
-package gyeongdan.article.repository;
+package gyeongdan.article.article.repository;
 
-import gyeongdan.article.domain.Article;
+import gyeongdan.article.article.domain.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
