+++ src/main/java/gyeongdan/article/article/repository/ArticleRepositoryImpl.java
@@ -1,10 +1,10 @@
-package gyeongdan.article.repository;
+package gyeongdan.article.article.repository;
 
-import gyeongdan.article.domain.Article;
+import gyeongdan.article.article.domain.Article;
 import java.util.List;
-import java.util.Optional;
+
+import gyeongdan.article.view_history.repository.ArticleViewHistoryJpaRepository;
 import lombok.AllArgsConstructor;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
