+++ src/main/java/gyeongdan/article/view_history/repository/ArticleViewHistoryJpaRepository.java
@@ -1,7 +1,7 @@
-package gyeongdan.article.repository;
+package gyeongdan.article.view_history.repository;
 
-import gyeongdan.article.domain.Article;
-import gyeongdan.article.domain.ArticleViewHistory;
+import gyeongdan.article.article.domain.Article;
+import gyeongdan.article.view_history.domain.ArticleViewHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
