+++ src/main/resources/application.properties
@@ -1,3 +1,3 @@
 spring.application.name=gyeongdan
-spring.config.import=optional:classpath:application.yml
+spring.config.import=optional:classpath:application-secret.yml
